+++ tests/unit/test_charm.py
@@ -19,6 +19,7 @@ from ops.model import ActiveStatus, BlockedStatus, MaintenanceStatus, StatusBase
 from ops.testing import Harness
 
 from charm import (
+    ACTIVE_STATUS_RECONCILIATION_FAILED_MSG,
     FAILED_RECONCILE_ACTION_ERR_MSG,
     GithubRunnerCharm,
     catch_action_errors,
@@ -307,7 +308,7 @@ def test_on_flush_runners_reconcile_error_fail(harness: Harness):
     """
     arrange: Set up charm with Openstack mode and ReconcileError.
     act: Run flush runner action.
-    assert: Action fails with generic message.
+    assert: Action fails with generic message and goes in ActiveStatus.
     """
     state_mock = MagicMock()
     state_mock.instance_type = InstanceType.OPENSTACK
@@ -320,6 +321,8 @@ def test_on_flush_runners_reconcile_error_fail(harness: Harness):
     mock_event = MagicMock()
     harness.charm._on_flush_runners_action(mock_event)
     mock_event.fail.assert_called_with(FAILED_RECONCILE_ACTION_ERR_MSG)
+    assert harness.charm.unit.status.name == ActiveStatus.name
+    assert harness.charm.unit.status.message == ACTIVE_STATUS_RECONCILIATION_FAILED_MSG
 
 
 def test_on_flush_runners_action_success(harness: Harness, runner_binary_path: Path):
@@ -340,7 +343,7 @@ def test_on_reconcile_runners_action_reconcile_error_fail(
     """
     arrange: Set up charm with Openstack mode, ready image, and ReconcileError.
     act: Run reconcile runners action.
-    assert: Action fails with generic message.
+    assert: Action fails with generic message and goes in ActiveStatus
     """
     state_mock = MagicMock()
     state_mock.instance_type = InstanceType.OPENSTACK
@@ -359,13 +362,15 @@ def test_on_reconcile_runners_action_reconcile_error_fail(
     harness.charm._on_reconcile_runners_action(mock_event)
 
     mock_event.fail.assert_called_with(FAILED_RECONCILE_ACTION_ERR_MSG)
+    assert harness.charm.unit.status.name == ActiveStatus.name
+    assert harness.charm.unit.status.message == ACTIVE_STATUS_RECONCILIATION_FAILED_MSG
 
 
 def test_on_reconcile_runners_reconcile_error(harness: Harness, monkeypatch: pytest.MonkeyPatch):
     """
     arrange: Set up charm with Openstack mode, ready image, and ReconcileError.
     act: Trigger reconcile_runners event.
-    assert: Unit goes into MaintenanceStatus with error message.
+    assert: Unit goes into ActiveStatus with error message.
     """
     state_mock = MagicMock()
     state_mock.instance_type = InstanceType.OPENSTACK
@@ -383,8 +388,8 @@ def test_on_reconcile_runners_reconcile_error(harness: Harness, monkeypatch: pyt
     mock_event = MagicMock()
     harness.charm._on_reconcile_runners(mock_event)
 
-    assert harness.charm.unit.status.name == MaintenanceStatus.name
-    assert harness.charm.unit.status.message == "Failed to reconcile runners"
+    assert harness.charm.unit.status.name == ActiveStatus.name
+    assert harness.charm.unit.status.message == ACTIVE_STATUS_RECONCILIATION_FAILED_MSG
 
 
 @pytest.mark.parametrize(